# api-docs/openAPI-specification.yml
openapi: 3.0.1
info:
  version: "1.0.0"
  title: RUMbus API
  description: >-
    RUMbus is a system to provide real-time information on the location of
    buses, as well as estimated arrival times at stops for the transportation
    service of the University of Puerto Rico at Mayaguez (UPRM).
...
paths:
...
  /users/{user_id}:
    put:
      tags:
        - Users
      summary: Update user by its id
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
          description: ID of the user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                email:
                  type: string
            examples:
              '0':
                value: |-
                  {
                      "full_name": "John Test"
                  }
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Users
      summary: Delete user by its id
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
          description: ID of the user
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Users
      summary: Get user by its id
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
          description: ID of the user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  full_name:
                    type: string
                  email:
                    type: string
                  updated_on:
                    type: string
                  created_on:
                    type: string
              examples:
                '0':
                  value: >-
                    {"id":"ee4c7483-504c-4ca6-80ff-453f4bfdf9ff","full_name":"John
                    Doe","email":"john.doe@test.com","updated_on":"2020-05-25T23:39:04.000Z","created_on":"2020-05-25T23:39:04.000Z"}
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"NotFound","message":"User with id:
                    ee4c7483-504c-4ca6-80ff-453f4bfdf9fa not found"}
        '500':
          $ref: '#/components/responses/500'
  /users/search:
    get:
      tags:
        - Users
      summary: Search for users with the specified parameters
      parameters:
        - name: id
          in: query
          schema:
            type: string
          example: User id
        - name: full_name
          in: query
          schema:
            type: string
          example: User full name
        - name: email
          in: query
          schema:
            type: string
          example: User email
        - name: skip
          in: query
          schema:
            type: string
          example: Parameter for pagination
        - name: limit
          in: query
          schema:
            type: string
          example: Parameter for pagination
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      type: string
                    full_name:
                      type: string
                    email:
                      type: string
                    created_on:
                      type: string
                    updated_on:
                      type: string
              examples:
                '/users/search?email=john.doe@test.com':
                  value: >-
                    [{"id":"f398fece-a30a-4004-bdd7-1d559db7cbcb","full_name":"John
                    Doe","email":"john.doe@test.com","updated_on":"2020-07-03T03:58:16.000Z","created_on":"2020-07-03T03:58:16.000Z"}]
        '500':
          $ref: '#/components/responses/500'
...
  /vehicles/{vehicle_id}/vehicle_maintenance:
    post:
      tags:
        - Vehicles
      parameters:
        - in: path
          name: vehicle_id
          schema:
            type: string
          required: true
          description: ID of the vehicle that was in maintenance
      summary: Create a maintenance record for the specified vehicle
      requestBody:
        content:
          application/json:
            schema:
              required: [cost, user_id, details, mileage, inspector, type, maintenance_date]
              type: object
              properties:
                cost:
                  type: number
                user_id:
                  type: string
                details:
                  type: string
                inspector:
                  type: string
                maintenance_date:
                  type: string
                type:
                  type: string
                mileage:
                  type: integer
            examples:
              '0':
                value: "{\r\n  \"user_id\": \"7da6943c-8ec0-4726-9168-40dd4b539af7\",\r\n  \"inspector\": \"Juan Test\",\r\n  \"cost\": 500.0,\r\n  \"mileage\": 130000,\r\n  \"details\": \"Broken parts\",\r\n  \"maintenance_date\": \"2019-03-05\",\r\n  \"type\": \"Fixes\"\r\n}"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: string
              examples:
                '0':
                  value: '"f2c97a48-8c98-4468-a958-5a5ecedd8bb7"'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide a vehicle
                    ID."}
                '1':
                  value: >-
                    {"code":"BadRequest","message":"Missing required param:
                    user_id"}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"NotFound","message":"Vehicle with id:
                    48a50fdb-9a2e-45f5-a0fa not found"}
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Vehicles
      parameters:
        - in: path
          name: vehicle_id
          schema:
            type: string
          required: true
          description: ID of the vehicle that was in maintenance
      summary: Get all maintenance records for the specified vehicle
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    cost:
                      type: number
                    user_id:
                      type: string
                    details:
                      type: string
                    inspector:
                      type: string
                    maintenance_date:
                      type: string
                    type:
                      type: string
                    mileage:
                      type: integer
                    created_on:
                      type: string
              examples:
                '0':
                  value: >-
                    [{"id":"83647f94-68bd-4d54-bf16-1a626a8a7cf1","user_id":"1","vehicle_id":"48a50fdb-9a2e-45f5-a0fa-57515155fcd9","mileage":130000,"type":"Arreglos","cost":500,"details":"Se
                    rompieron piezas","inspector":"Juan
                    Test","maintenance_date":"2019-03-05T00:00:00.000Z","created_on":"2020-03-22T21:27:53.000Z"},{"id":"abcd7a48-8c98-4468-a958-5a5ecedd8bb7","user_id":"1234","vehicle_id":"48a50fdb-9a2e-45f5-a0fa-57515155fcd9","mileage":130000,"type":"Arreglos","cost":500,"details":"Se
                    rompieron piezas","inspector":"Juan
                    Test","maintenance_date":"2019-03-05T00:00:00.000Z","created_on":"2020-03-22T21:29:38.000Z"}]
                '1':
                  value: >-
                    []
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide a vehicle
                    ID."}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"NotFound","message":"Vehicle with id:
                    48a50fdb-9a2e-45f5 not found"}
        '500':
          $ref: '#/components/responses/500'
  /vehicles/vehicle_maintenance/{maintenance_id}:
    put:
      tags:
        - Vehicles
      summary: Update a vehicle maintenance record
      parameters:
        - in: path
          name: maintenance_id
          schema:
            type: string
          required: true
          description: ID of the vehicle maintenance record
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cost:
                  type: number
                user_id:
                  type: string
                details:
                  type: string
                inspector:
                  type: string
                maintenance_date:
                  type: string
                type:
                  type: string
                mileage:
                  type: integer
            examples:
              '0':
                value: |-
                  {
                      "cost": 1000
                  }
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'
...
components:
  responses:
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
          examples:
            'Server Error':
              value: '{"code":"InternalServer","message":"An internal server error ocurred"}'